+++ packages/react-dom/src/__tests__/ReactDOMServerPartialHydration-test.internal.js
@@ -85,7 +85,7 @@ describe('ReactDOMServerPartialHydration', () => {
     ReactDOMServer = require('react-dom/server');
     Scheduler = require('scheduler');
     Suspense = React.Suspense;
-    SuspenseList = React.unstable_SuspenseList;
+    SuspenseList = React.SuspenseList;
 
     useHover = require('react-interactions/events/hover').useHover;
   });
@@ -344,7 +344,7 @@ describe('ReactDOMServerPartialHydration', () => {
     }).toWarnDev(
       'Warning: Cannot hydrate Suspense in legacy mode. Switch from ' +
         'ReactDOM.hydrate(element, container) to ' +
-        'ReactDOM.unstable_createSyncRoot(container, { hydrate: true })' +
+        'ReactDOM.createBlockingRoot(container, { hydrate: true })' +
         '.render(element) or remove the Suspense components from the server ' +
         'rendered components.' +
         '\n    in Suspense (at **)' +