+++ packages/react-dom/src/__tests__/ReactDOMServerSuspense-test.internal.js
@@ -107,14 +107,14 @@ describe('ReactDOMServerSuspense', () => {
 
   it('server renders a SuspenseList component and its children', async () => {
     const example = (
-      <React.unstable_SuspenseList>
+      <React.SuspenseList>
         <React.Suspense fallback="Loading A">
           <div>A</div>
         </React.Suspense>
         <React.Suspense fallback="Loading B">
           <div>B</div>
         </React.Suspense>
-      </React.unstable_SuspenseList>
+      </React.SuspenseList>
     );
     const element = await serverRender(example);
     const parent = element.parentNode;
@@ -126,7 +126,7 @@ describe('ReactDOMServerSuspense', () => {
     expect(divB.textContent).toBe('B');
 
     ReactTestUtils.act(() => {
-      const root = ReactDOM.createSyncRoot(parent, {hydrate: true});
+      const root = ReactDOM.createBlockingRoot(parent, {hydrate: true});
       root.render(example);
     });