+++ packages/react-dom/src/server/ReactDOMFizzServerBrowser.js
@@ -13,7 +13,7 @@ import {
   createRequest,
   startWork,
   startFlowing,
-} from 'react-stream/inline.dom-browser';
+} from 'react-server/inline.dom-browser';
 
 function renderToReadableStream(children: ReactNodeList): ReadableStream {
   let request;