+++ packages/shared/ReactFeatureFlags.js
@@ -69,7 +69,7 @@ export const enableScopeAPI = false;
 export const enableJSXTransformAPI = false;
 
 // We will enforce mocking scheduler with scheduler/unstable_mock at some point. (v17?)
-// Till then, we warn about the missing mock, but still fallback to a sync mode compatible version
+// Till then, we warn about the missing mock, but still fallback to a legacy mode compatible version
 export const warnAboutUnmockedScheduler = false;
 
 // For tests, we flush suspense fallbacks in an act scope;